@@ -3,15 +3,13 @@
   "description": "create droste images with droste effect",
   "version": "1.0.0",
   "private": true,
-  "homepage": "https://javier.xyz/droste-creator/",
+  "homepage": "/droste-creator/",
   "dependencies": {
-    "capsize": "^2.0.0",
-    "jbx": "^1.7.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "jbx": "^2.4.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-draggable": "^4.4.4",
-    "react-scripts": "^5.0.0",
-    "styled-components": "^5.3.3"
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "dev": "react-scripts start",
@@ -47,6 +45,6 @@
   "devDependencies": {
     "eslint": "^8.8.0",
     "eslint-plugin-react": "^7.28.0",
-    "gh-pages": "^3.2.3"
+    "gh-pages": "^4.0.0"
   }
 }
