@@ -1,5 +1,9 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './app';
+import './style.css';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './app.jsx';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(<App />);
